backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
asgiref==3.8.1
sqlparse==0.5.3 